Better usage of CUDA math functions
There are overload of math functions, using them will ensure the
computation will always be performed correctly

+++ src/log.cu
@@ -13,7 +13,7 @@ __global__ void log_kernel(size_t n, T alpha, const T* x, size_t incx, T* y, siz
     auto stride = blockDim.x * gridDim.x;
 
     for (; index < n; index += stride) {
-        y[incy * index] = alpha * logf(x[incx * index]);
+        y[incy * index] = alpha * log(x[incx * index]);
     }
 }
 
@@ -26,10 +26,9 @@ __global__ void log_kernel(size_t n, cuComplex alpha, const cuComplex* x, size_t
         cuComplex c = x[incx * index];
 
         float c_abs = hypot(c.x, c.y);
-        float c_arg = atan2f(c.y, c.x);
+        float c_arg = atan2(c.y, c.x);
 
-        y[incx * index] = cuCmulf(alpha, make_cuComplex(logf(c_abs), c_arg));
-
+        y[incx * index] = cuCmulf(alpha, make_cuComplex(log(c_abs), c_arg));
     }
 }
 
@@ -42,7 +41,7 @@ __global__ void log_kernel(size_t n, cuDoubleComplex alpha, const cuDoubleComple
         cuDoubleComplex c = x[incx * index];
 
         double c_abs = hypot(c.x, c.y);
-        double c_arg = atan2f(c.y, c.x);
+        double c_arg = atan2(c.y, c.x);
 
         y[incx * index] = cuCmul(alpha, make_cuDoubleComplex(log(c_abs), c_arg));
     }
@@ -54,7 +53,7 @@ __global__ void log_kernel1(size_t n, const T* x, size_t incx, T* y, size_t incy
     auto stride = blockDim.x * gridDim.x;
 
     for (; index < n; index += stride) {
-        y[incy * index] = logf(x[incx * index]);
+        y[incy * index] = log(x[incx * index]);
     }
 }
 
@@ -67,9 +66,9 @@ __global__ void log_kernel1(size_t n, const cuComplex* x, size_t incx, cuComplex
         cuComplex c = x[incx * index];
 
         float c_abs = hypot(c.x, c.y);
-        float c_arg = atan2f(c.y, c.x);
+        float c_arg = atan2(c.y, c.x);
 
-        y[incx * index] = make_cuComplex(logf(c_abs), c_arg);
+        y[incx * index] = make_cuComplex(log(c_abs), c_arg);
     }
 }
 
@@ -82,7 +81,7 @@ __global__ void log_kernel1(size_t n, const cuDoubleComplex* x, size_t incx, cuD
         cuDoubleComplex c = x[incx * index];
 
         double c_abs = hypot(c.x, c.y);
-        double c_arg = atan2f(c.y, c.x);
+        double c_arg = atan2(c.y, c.x);
 
         y[incx * index] = make_cuDoubleComplex(log(c_abs), c_arg);
     }
